% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagfLasso.R
\name{oracle}
\alias{oracle}
\title{Grouped Panel Data Oracle Estimator}
\usage{
oracle(
  y,
  X,
  n_periods,
  groups_0,
  method = "PLS",
  Z = NULL,
  bias_correc = FALSE
)
}
\arguments{
\item{y}{a \eqn{NT \times 1} vector or data.frame of the dependent variable, with \eqn{\bold{y}=(y_1, \dots, y_N)^\prime}, \eqn{y_i = (y_{i1}, \dots, y_{iT})^\prime} and the scalar \eqn{y_{it}}.}

\item{X}{a \eqn{NT \times p} matrix or data.frame of explanatory variables, with \eqn{\bold{X}=(x_1, \dots, x_N)^\prime}, \eqn{x_i = (x_{i1}, \dots, x_{iT})^\prime} and the \eqn{p \times 1} vector \eqn{x_{it}}.}

\item{n_periods}{the number of observed time periods \eqn{T}.}

\item{groups_0}{the true group vector \eqn{(g_1, \dots, g_N)}, where \eqn{g_i = k} if the unit \eqn{i} is part of group \eqn{k}.}

\item{method}{the estimation method. Options are
\describe{
\item{\code{'PLS'}}{for using the penalized least squares (\emph{PLS}) algorithm. We recommend \emph{PLS} in case of (weakly) exogenous regressors (Mehrabani, 2023, sec. 2.2).}
\item{\code{'PGMM'}}{for using the penalized Generalized Method of Moments (\emph{PGMM}). \emph{PGMM} is required when instrumenting endogenous regressors (Mehrabani, 2023, sec. 2.3). A matrix \eqn{Z} contains the necessary exogenous instruments.}
} Default is \code{'PLS'}.}

\item{Z}{a \eqn{NT \times q} matrix of exogenous instruments, where \eqn{q \geq p}, \eqn{\bold{Z}=(z_1, \dots, z_N)^\prime}, \eqn{z_i = (z_{i1}, \dots, z_{iT})^\prime} and \eqn{z_{it}} is a \eqn{q \times 1} vector. \eqn{Z} is only required when \code{method = 'PGMM'} is selected. When using \code{'PLS'}, either pass \code{NULL} or any matrix \eqn{\bold{Z}} is disregarded. Default is \code{NULL}.}

\item{bias_correc}{logical. If \code{TRUE}, a Split-panel Jackknife bias correction following Dhaene and Jochmans (2015) is applied to the slope parameters. We recommend using this correction when facing a dynamic panel. Default is \code{FALSE}.}
}
\value{
A \eqn{K \times p} matrix of the group-specific slope coefficients.
}
\description{
Estimation of a panel data model, given a known group structure.
}
\details{
Given a known group pattern \eqn{\bold{G}^0}, the least squares oracle estimator (pendant to \emph{PLS}) solves
\deqn{\hat{\alpha}_k^{\text{oracle}} = \left( \sum_{i \in G_k^0} \sum_{t = 1}^T \tilde{x}_{it} \tilde{x}_{it}^\prime \right)^{-1} \sum_{i \in G_k^0} \sum_{t = 1}^T \tilde{x}_{it} \tilde{y}_{it},}
where \eqn{\tilde{y}_{it}} is the demeaned dependent variable and \eqn{\tilde{x}_{it}} represents a vector of demeaned weakly exogenous explanatory variables.

The oracle estimates equivalent to \emph{PGMM} are obtained by
\deqn{\hat{\alpha}_k^{\text{oracle}} = \left( \tilde{Q}_{k \Delta x}^{(k)^\prime} \tilde{W}_{NT}^{(k)} \tilde{Q}_{k \Delta x}^{(k)} \right)^{-1} \tilde{Q}_{k \Delta x}^{(k)^\prime} \tilde{W}_{NT}^{(k)} \tilde{Q}_{k \Delta y}^{(k)},}
where \eqn{\tilde{Q}_{k \Delta x}^{(k)} = T^{-1} \sum_{i \in G_k^0} \sum_{t = 1}^T z_{it} (\Delta x_{it})^\prime}, \eqn{\tilde{Q}_{k \Delta y}^{(k)} = T^{-1} \sum_{i \in G_k^0} \sum_{t = 1}^T z_{it} \Delta y_{it}}, and \eqn{\tilde{W}_{NT}^{(k)}} is a matrix containing weights. \eqn{z_{it}} represents a \eqn{q \times 1} vector of exogenous instruments employed to control for the endogenous \eqn{x_{it}}.
}
\examples{
# Simulate a panel with a latent group pattern
sim <- sim_DGP(N = 50, n_periods = 80, p = 2, n_groups = 3)
y <- sim$y
X <- sim$X
groups_0 <- sim$groups

# Estimate the grouped panel data model model, given the true grouping
estim <- oracle(y = y, X = X, n_periods = 80, groups_0 = groups_0)
print(estim)
}
\author{
Paul Haimerl
}
