% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim.R
\name{sim_DGP}
\alias{sim_DGP}
\title{Simulate a Panel With a Latent Group Structure}
\usage{
sim_DGP(
  N = 50,
  n_periods = 40,
  p = 2,
  n_groups = max(2, floor(N/3)),
  group_proportions = NULL,
  error_spec = NULL,
  dyn_panel = FALSE,
  q = NULL,
  alpha_0 = NULL
)
}
\arguments{
\item{N}{the number of cross-sectional units.}

\item{n_periods}{the number of simulated time periods \eqn{T}.}

\item{p}{the number of explanatory variables.}

\item{n_groups}{the number of latent groups \eqn{K}.}

\item{group_proportions}{a numeric vector of length \code{n_groups} indicating the fraction of \eqn{N} each group will contain. If \code{NULL}, all groups are of size \eqn{\frac{N}{K}}. Default is \code{NULL}.}

\item{error_spec}{the error specification used. Options are
\describe{
\item{\code{NULL}}{for \eqn{iid} errors.}
\item{\code{'AR'}}{for an \eqn{AR(1)} error process with an autoregressive coefficient of 0.5.}
\item{\code{'GARCH'}}{for a \eqn{GARCH(1,1)} error process with a 0.05 constant, a 0.05 ARCH and a 0.9 GARCH coefficient.}
}
Default is \code{NULL}.}

\item{dyn_panel}{Logical. If \code{TRUE}, the panel includes one stationary autoregressive lag of the dependent variable (see sec. Details for information on the \eqn{AR} coefficient). Default is \code{FALSE}.}

\item{q}{the number of simulated exogenous instruments when a panel with endogenous regressors is generated. If \code{NULL}, the regressors are exogenous. Default is \code{NULL}.}

\item{alpha_0}{optional pre-specified \eqn{K \times p} parameter matrix. If \code{NULL}, the coefficients are drawn randomly (see sec. Details). If \code{dyn_panel = TRUE}, the first column represents the stationary \eqn{AR} coefficient. Default is \code{NULL}.}
}
\value{
A list holding
\item{\code{alpha}}{the \eqn{K \times p} matrix of group-specific slope parameters. In case of \code{dyn_panel = TRUE}, the first column holds the \eqn{AR} coefficient.}
\item{\code{groups}}{a vector indicating the group memberships.}
\item{\code{y}}{a \eqn{NT \times 1} vector of the dependent variable, with \eqn{y=(y_1, \dots, y_N)^\prime}, \eqn{y_i = (y_{i1}, \dots, y_{iT})^\prime} and the scalar \eqn{y_{it}}.}
\item{\code{X}}{a \eqn{NT \times p} matrix of explanatory variables, with \eqn{X=(x_1, \dots, x_N)^\prime}, \eqn{x_i = (x_{i1}, \dots, x_{iT})^\prime} and the \eqn{p \times 1} vector \eqn{x_{it}}.}
\item{\code{Z}}{a \eqn{NT \times q} matrix of instruments , where \eqn{q \geq p}, \eqn{Z=(z_1, \dots, z_N)^\prime}, \eqn{z_i = (z_{i1}, \dots, z_{iT})^\prime} and \eqn{z_{it}} is a \eqn{q \times 1} vector. In case a panel with exogenous regressors is generated (\code{q = NULL}), \eqn{Z} equals \code{NULL}.}
}
\description{
Construct a static or dynamic, exogenous or endogenous panel data set subject to a latent group structure with optional \eqn{AR(1)} or \eqn{GARCH(1,1)} innovations.
}
\details{
The function nests the DGPs employed in the simulation study of Mehrabani (2023, sec. 6).

The \eqn{p} exogenous regressors are generated as follows:
\deqn{x_{it,j} = 0.2 \eta_i + e_{it,j}, \quad \eta_i,e_{it,j} \sim i.i.d. N(0, 1), \quad j = \{1, \dots, p\},}
where \eqn{\eta_i} represents individual fixed effects. \eqn{\eta_i} and \eqn{e_i} are independent of each other.

In case \code{alpha_0 = NULL}, the group-specific slope parameters \eqn{\alpha_{kj}} are drawn from \eqn{\sim U[-2, 2]}.

If a dynamic panel is specified (\code{dyn_panel = TRUE}), the \eqn{AR} coefficients \eqn{\beta^{\text{AR}}_i} are drawn from a uniform distribution with support \eqn{(-1, 1)} and \eqn{x_{it,j} = e_{it,j}}.
The individual fixed effects enter the dependent variable via \eqn{(1 - \beta^{\text{AR}}_i) \eta_i} to account for the autoregressive dependency (note that \eqn{\alpha_k} refers to group-specific and \eqn{\beta_i} to individual parameter values).
I refer to Mehrabani (2023, sec 6) for details.

When specifying an endogenous panel (set \code{q} to \eqn{\geq p}), \eqn{e_{it,j}} correlate with the cross-sectional innovations \eqn{u_{it}} by a magnitude of 0.5 to produce endogenous regressors with \eqn{E(u|X) \neq 0}. However, the endogenous regressors can be accounted for by exploiting the \eqn{q} instruments in \eqn{Z}, for which \eqn{E(u|Z) = 0} holds.
The instruments and the first stage coefficients are generated in the same fashion as \eqn{X} (\eqn{X} when \code{q = NULL}) and as \eqn{\alpha}, respectively.
}
\examples{
# Simulate DGP 1 from Mehrabani (2023, sec. 6)
alpha_0_DGP1 <- matrix(c(0.4, 1, 1.6, 1.6, 1, 0.4), ncol = 2)
DGP1 <- sim_DGP(
  N = 50, n_periods = 20, p = 2, n_groups = 3,
  group_proportions = c(.4, .3, .3), alpha_0 = alpha_0_DGP1
)

# Simulate DGP 6 from Mehrabani (2023, sec. 6)
alpha_0_DGP6 <- cbind(
  c(0.8, 0.6, 0.4, 0.2, -0.2, -0.4, -0.6, -0.8),
  c(-4, -3, -2, -1, 1, 2, 3, 4),
  c(4, 3, 2, 1, -1, -2, -3, -4)
)
DGP6 <- sim_DGP(
  N = 50, n_periods = 20, p = 2, n_groups = 8,
  group_proportions = c(.3, rep(.1, 7)), dyn_panel = TRUE,
  alpha_0 = alpha_0_DGP6
)
}
\references{
Mehrabani, A. (2023). Estimation and identification of latent group structures in panel data. \emph{Journal of Econometrics}, 235(2), 1464-1482. \doi{10.1016/j.jeconom.2022.12.002}.
}
\author{
Paul Haimerl
}
