% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynpagfLasso.R
\name{tv_oracle}
\alias{tv_oracle}
\title{Grouped Time-varying Panel Data Oracle Estimator}
\usage{
tv_oracle(y, X, n_periods, groups_0, d = 3, J = floor(NROW(y)^(1/7)))
}
\arguments{
\item{y}{a \eqn{NT \times 1} vector or data.frame of the dependent variable, with \eqn{\bold{y}=(y_1, \dots, y_N)^\prime}, \eqn{y_i = (y_{i1}, \dots, y_{iT})^\prime} and the scalar \eqn{y_{it}}.}

\item{X}{a \eqn{NT \times p} matrix or data.frame of explanatory variables, with \eqn{\bold{X}=(x_1, \dots, x_N)^\prime}, \eqn{x_i = (x_{i1}, \dots, x_{iT})^\prime} and the \eqn{p \times 1} vector \eqn{x_{it}}.}

\item{n_periods}{the number of observed time periods \eqn{T}.}

\item{groups_0}{the true group vector \eqn{(g_1, \dots, g_N)}, where \eqn{g_i = k} if the unit \eqn{i} is part of group \eqn{k}.}

\item{d}{the degree of the coefficient polynomial spline functions. Default is 3.}

\item{J}{The number of interior knots of the spline functions. If left unspecified, the default heuristic \eqn{\text{floor}((NT)^{\frac{1}{7}})} is used.}
}
\value{
A \eqn{T \times p \times K} array of the group-specific time-varying coefficients.
}
\description{
Estimation of a time-varying panel data model, given a known group structure.
}
\details{
Given a known group pattern \eqn{\bold{G}^0}, the least squares oracle estimator solves
\deqn{\text{vec}(\hat{\xi}_k^{\text{oracle}}) = \left( \sum_{i \in G_k^0} \sum_{t = 1}^T \tilde{z}_{it} \tilde{z}_{it}^\prime \right)^{-1} \sum_{i \in G_k^0} \sum_{t = 1}^T \tilde{z}_{it} \tilde{y}_{it},}
where \eqn{\tilde{y}_{it}} is the demeaned dependent variable and \eqn{\tilde{z}_{it}} is a \eqn{(J + d + 1) \times 1} vector that interacts the explanatory variables \eqn{x_{it}} with the spline basis system of degree \eqn{d} and \eqn{J} interior knots \eqn{\bold{B}(v)}, demeaned.
The final time-varying coefficients are obtained by \eqn{\hat{\alpha}_{kt}^{\text{oracle}} = \hat{\xi}_k^{\text{oracle}^\prime} \bold{B} \left( \frac{t}{T} \right)}.
}
\examples{
# Simulate a time-varying panel with a trend and a group pattern
sim <- sim_tv_DGP(N = 20, n_periods = 50, DGP = 1)
y <- sim$y
X <- sim$X
groups_0 <- sim$groups

# Estimate the time-varying grouped panel data model model, given the true grouping
estim <- tv_oracle(y = y, X = X, n_periods = 50, groups_0 = groups_0)
print(estim)
}
\author{
Paul Haimerl
}
