% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynSim.R
\name{sim_tv_DGP}
\alias{sim_tv_DGP}
\title{Simulate a Time-varying Panel With a Latent Group Structure}
\usage{
sim_tv_DGP(N = 50, n_periods = 40, DGP = 1, sd_error = 1)
}
\arguments{
\item{N}{the number of cross-sectional units. Default is 50.}

\item{n_periods}{the number of simulated time periods \eqn{T}. Default is 40.}

\item{DGP}{the data generating process. Options are
\describe{
\item{1}{generates a trend only, according to DGP 1 in Su et al. (2019, sec. 6.1)}
\item{2}{simulates a trend and an additional exogenous explanatory variable, as described by DGP 2 in Su et al. (2019, sec. 6.1)}
\item{1}{draws a dynamic panel data model with one \eqn{AR} lag, following DGP 3 in Su et al. (2019, sec. 6.1)}
}}

\item{sd_error}{standard deviation of the cross-sectional errors. Default is 1.}
}
\value{
A list holding
\item{\code{alpha}}{a \eqn{T \times p \times K} array of group-specific time-varying parameters}
\item{\code{beta}}{a \eqn{T \times p \times N} array of individual time-varying parameters}
\item{\code{groups}}{a vector indicating the group memberships \eqn{(g_1, \dots, g_N)}, where \eqn{g_i = k} if \eqn{i \in} group \eqn{k}.}
\item{\code{y}}{a \eqn{NT \times 1} vector of the dependent variable, with \eqn{\bold{y}=(y_1, \dots, y_N)^\prime}, \eqn{y_i = (y_{i1}, \dots, y_{iT})^\prime} and the scalar \eqn{y_{it}}.}
\item{\code{X}}{a \eqn{NT \times p} matrix of explanatory variables, with \eqn{\bold{X}=(x_1, \dots, x_N)^\prime}, \eqn{x_i = (x_{i1}, \dots, x_{iT})^\prime} and the \eqn{p \times 1} vector \eqn{x_{it}}.}
}
\description{
Construct a time-varying panel data set subject to a latent group structure with three distinct groups, mirroring Su et al. (2019, sec. 6.1).
}
\details{
The scalar dependent variable \eqn{y_{it}} is driven by the following panel data model:
\deqn{y_{it} = \gamma_i + \beta^\prime_{it} x_{it} + u_{it}, \quad i = 1, \dots, N, \; t = 1, \dots, T,}
where \eqn{y_{it}} is the scalar dependent variable, \eqn{\gamma_i} is an individual fixed effect and \eqn{x_{it}} is a \eqn{p \times 1} vector of explanatory variables. The errors \eqn{u_{it}} feature a \eqn{iid} standard normal distribution.
The coefficient vector \eqn{\beta_i = \{\beta_{i1}^\prime, \dots, \beta_{iT}^\prime \}^\prime} is subject to the group pattern
\deqn{\beta_i = \sum_{k = 1}^3 \alpha_k \bold{1} \{i \in G_k \},}
with \eqn{\cup_{k = 1}^3 G_k = \{1, \dots, N\}}, \eqn{G_k \cap G_j = \emptyset} and \eqn{\| \alpha_k \| \neq \| \alpha_j \|} for any \eqn{k \neq j}.

\describe{
\item{\code{DGP = 1}}{\eqn{x_{it} = 1} and thus only includes a group-specific trend.}
\item{\code{DGP = 2}}{\eqn{x_{it} = (1, x_{it,2})^\prime}, where \eqn{x_{it,2}} is standard normal.}
\item{\code{DGP = 3}}{\eqn{x_{it} = y_{it-1}}.}
}

The parameterization of the functional coefficients and group sizes follow Su et al. (2019, sec. 6.1).
}
\examples{
# Simulate a time-varying panel subject to a time trend and a latent group structure
sim <- sim_tv_DGP(N = 20, n_periods = 50, DGP = 1)
y <- sim$y
X <- sim$X
}
\references{
Su, L., Wang, X., & Jin, S. (2019). Sieve estimation of time-varying panel data models with latent structures. \emph{Journal of Business & Economic Statistics}, 37(2), 334-349. \doi{10.1080/07350015.2017.1340299}.
}
\author{
Paul Haimerl
}
